remove type casting Double due to compilation error

+++ scenario/src/main/java/gov/nasa/arc/mct/scenario/component/ActivityTypeModel.java
@@ -59,7 +59,7 @@ public class ActivityTypeModel {
 		if (costs.get(typeName) == null) {
 			return 0.0;
 		} else {
-			return (Double)costs.get(typeName);
+			return costs.get(typeName);
 		}
 	}
